# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sovereign Agent Cloud",
    "version": "0.1.0",
    "description": "All browser POST requests require an Origin matching PUBLIC_ORIGIN. Wallet values are integer micro-euro strings. Session cookie is HttpOnly/Secure/SameSite=Strict. Upstream provider keys never enter this API."
  },
  "servers": [
    {
      "url": "https://account.example.com"
    }
  ],
  "paths": {
...
    "/api/auth/request": {
      "post": {
        "summary": "Send verified-email sign-in link",
        "operationId": "post_api_auth_request",
        "responses": {
          "200": {
            "description": "Successful JSON response; URL responses contain a url field."
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin, ownership, or fresh login check failed"
          },
          "409": {
            "description": "Lifecycle conflict"
          },
          "503": {
            "description": "Launch gate closed or dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/auth/consume": {
      "post": {
        "summary": "Consume a single-use sign-in token",
        "operationId": "post_api_auth_consume",
        "responses": {
          "200": {
            "description": "Successful JSON response; URL responses contain a url field."
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin, ownership, or fresh login check failed"
          },
          "409": {
            "description": "Lifecycle conflict"
          },
          "503": {
            "description": "Launch gate closed or dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "End the account session",
        "operationId": "post_api_auth_logout",
        "responses": {
          "200": {
            "description": "Successful JSON response; URL responses contain a url field."
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Origin, ownership, or fresh login check failed"
          },
          "409": {
            "description": "Lifecycle conflict"
          },
          "503": {
            "description": "Launch gate closed or dependency unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      },
      "tenantKey": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}